Namespace corrected for dmlf ditributed learning and moved utilities to reduce reusing
