+++ libs/dmlf/examples/boston_housing_multiprocess_distributed_learning/main.cpp
@@ -17,6 +17,8 @@
 //------------------------------------------------------------------------------
 
 #include "dmlf/distributed_learning/distributed_learning_client.hpp"
+#include "dmlf/distributed_learning/utilities/boston_housing_client_utilities.hpp"
+#include "dmlf/distributed_learning/utilities/distributed_learning_utilities.hpp"
 #include "dmlf/networkers/muddle_learner_networker.hpp"
 #include "dmlf/simple_cycling_algorithm.hpp"
 #include "math/matrix_operations.hpp"
@@ -26,7 +28,6 @@
 #include "ml/exceptions/exceptions.hpp"
 #include "ml/ops/loss_functions/cross_entropy_loss.hpp"
 #include "ml/optimisation/adam_optimiser.hpp"
-#include "ml/utilities/boston_housing_client_utilities.hpp"
 
 #include <algorithm>
 #include <chrono>
@@ -39,71 +40,12 @@
 
 using namespace fetch::ml::ops;
 using namespace fetch::ml::layers;
-using namespace fetch::ml::distributed_learning;
+using namespace fetch::dmlf::distributed_learning;
 
 using DataType         = fetch::fixed_point::FixedPoint<32, 32>;
 using TensorType       = fetch::math::Tensor<DataType>;
 using VectorTensorType = std::vector<TensorType>;
 using SizeType         = fetch::math::SizeType;
-
-/**
- * Get loss of given model on given dataset
- * @param g_ptr model
- * @param data_tensor input
- * @param label_tensor label
- * @return
- */
-DataType Test(std::shared_ptr<fetch::ml::Graph<TensorType>> const &g_ptr,
-              TensorType const &data_tensor, TensorType const &label_tensor)
-{
-  g_ptr->SetInput("Input", data_tensor);
-  g_ptr->SetInput("Label", label_tensor);
-  return *(g_ptr->Evaluate("Error").begin());
-}
-
-void Shuffle(TensorType &data, TensorType &labels, SizeType const &seed = 54)
-{
-  TensorType data_out   = data.Copy();
-  TensorType labels_out = labels.Copy();
-
-  std::vector<SizeType> indices;
-  SizeType              axis = data.shape().size() - 1;
-
-  for (SizeType i{0}; i < data.shape().at(axis); i++)
-  {
-    indices.push_back(i);
-  }
-
-  fetch::random::LaggedFibonacciGenerator<> lfg(seed);
-  fetch::random::Shuffle(lfg, indices, indices);
-
-  for (SizeType i{0}; i < data.shape().at(axis); i++)
-  {
-    auto data_it       = data.View(i).begin();
-    auto data_out_it   = data_out.View(indices.at(i)).begin();
-    auto labels_it     = labels.View(i).begin();
-    auto labels_out_it = labels_out.View(indices.at(i)).begin();
-
-    while (data_it.is_valid())
-    {
-      *data_out_it = *data_it;
-
-      ++data_it;
-      ++data_out_it;
-    }
-
-    while (labels_it.is_valid())
-    {
-      *labels_out_it = *labels_it;
-
-      ++labels_it;
-      ++labels_out_it;
-    }
-  }
-
-  data   = data_out;
-  labels = labels_out;
-}
 
 int main(int argc, char **argv)
 {
@@ -142,9 +84,9 @@ int main(int argc, char **argv)
       fetch::ml::dataloaders::ReadCSV<TensorType>(labels_filename).Transpose();
 
   // Shuffle data
-  Shuffle(data_tensor, label_tensor, seed);
+  utilities::Shuffle(data_tensor, label_tensor, seed);
 
-  auto client = fetch::ml::utilities::MakeBostonClient<TensorType>(
+  auto client = fetch::dmlf::distributed_learning::utilities::MakeBostonClient<TensorType>(
       std::to_string(instance_number), client_params, data_tensor, label_tensor, test_set_ratio,
       console_mutex_ptr);
 
@@ -172,10 +114,10 @@ int main(int argc, char **argv)
     client->Run();
 
     std::cout << it << "\t"
-              << static_cast<double>(Test(client->GetModel(), data_tensor, label_tensor))
+              << static_cast<double>(utilities::Test(client->GetModel(), data_tensor, label_tensor))
               << std::endl;
     lossfile << it << ","
-             << static_cast<double>(Test(client->GetModel(), data_tensor, label_tensor))
+             << static_cast<double>(utilities::Test(client->GetModel(), data_tensor, label_tensor))
              << std::endl;
   }
