+++ libs/ml/include/ml/core/graph.hpp
@@ -35,6 +35,14 @@
 // appear in middle of graph
 
 namespace fetch {
+
+namespace dmlf {
+namespace distributed_learning {
+template <typename TensorType>
+class TrainingClient;
+}  // namespace distributed_learning
+}  // namespace dmlf
+
 namespace ml {
 
 ///////////////
@@ -50,11 +58,6 @@ namespace model {
 template <typename TensorType>
 class ModelInterface;
 }  // namespace model
-
-namespace distributed_learning {
-template <typename TensorType>
-class TrainingClient;
-}  // namespace distributed_learning
 
 ///////////////////
 /// GRAPH STATE ///
@@ -165,7 +168,7 @@ private:
 
   friend class optimisers::Optimiser<TensorType>;
   friend class model::ModelInterface<TensorType>;
-  friend class distributed_learning::TrainingClient<TensorType>;
+  friend class dmlf::distributed_learning::TrainingClient<TensorType>;
 
   TensorType ForwardImplementation(std::string const &node_name, bool is_training,
                                    bool evaluate_mode);
